change counter int type to int64

+++ metacounterdriver.go
@@ -5,7 +5,7 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type meta struct {
 	ID     primitive.ObjectID
 	Meta   string
-	Amount int
+	Amount int64
 }
 
 type metaCounter struct {
@@ -18,7 +18,7 @@ func (mc *metaCounter) addCol(col string) {
 	}
 }
 
-func (mc *metaCounter) Add(_col string, _meta string, id *primitive.ObjectID, amount int) MetaCounter {
+func (mc *metaCounter) Add(_col string, _meta string, id *primitive.ObjectID, amount int64) MetaCounter {
 	if id != nil {
 		mc.addCol(_col)
 		for i, mt := range mc.Data[_col] {
@@ -34,14 +34,14 @@ func (mc *metaCounter) Add(_col string, _meta string, id *primitive.ObjectID, am
 
 func (mc *metaCounter) Build() []MetaCounterResult {
 	result := make([]MetaCounterResult, 0)
-	ignores := make(map[string]map[string]int)
-	addIgnore := func(_col, _meta string, amount int) {
+	ignores := make(map[string]map[string]int64)
+	addIgnore := func(_col, _meta string, amount int64) {
 		if _, ok := ignores[_col]; !ok {
-			ignores[_col] = make(map[string]int)
+			ignores[_col] = make(map[string]int64)
 		}
 		ignores[_col][_meta] = amount
 	}
-	isAdded := func(_col, _meta string, amount int) bool {
+	isAdded := func(_col, _meta string, amount int64) bool {
 		for k, i := range ignores {
 			if k == _col {
 				for _k, _a := range i {
@@ -53,7 +53,7 @@ func (mc *metaCounter) Build() []MetaCounterResult {
 		}
 		return false
 	}
-	foundIds := func(_meta string, amount int, data []meta) []primitive.ObjectID {
+	foundIds := func(_meta string, amount int64, data []meta) []primitive.ObjectID {
 		ids := make([]primitive.ObjectID, 0)
 		for _, m := range data {
 			if m.Meta == _meta && amount == m.Amount {
@@ -68,7 +68,7 @@ func (mc *metaCounter) Build() []MetaCounterResult {
 				result = append(result, MetaCounterResult{
 					Col:    _col,
 					Ids:    foundIds(m.Meta, m.Amount, _meta),
-					Values: map[string]int{m.Meta: m.Amount},
+					Values: map[string]int64{m.Meta: m.Amount},
 				})
 				addIgnore(_col, m.Meta, m.Amount)
 			}
